Chore: Move default encryption provider to kmsproviders package
